handle None case

+++ h2integrate/core/h2integrate_model.py
@@ -1332,7 +1332,10 @@ class H2IntegrateModel:
             self.prob.setup()
             self.setup_has_been_called = True
 
-        self.recorder_path.unlink(missing_ok=True)
+        # OpenMDAO will skip this step if it encounters an issue leading to silent failures
+        if self.recorder_path is not None:
+            self.recorder_path.unlink(missing_ok=True)
+
         self.prob.run_driver()
 
     def post_process(self, print_results=True, summarize_sql=False, show_plots=False):
